requirements_raspberry.txt
pigpio==1.78
RPi.GPIO==0.7.1
smbus2==0.6.1
